Refactored to use json file rather than ts for source of articles
