--- app/blog/article.ts
@@ -1,76 +1,0 @@
-// articles.js
-const articles = [
-    {
-      title: "V3 Model: New Records!",
-      image: "/v3_w_background.jpg",
-      author: "Tran Ngoc Duy Ngo",
-      date: "XX-ABC-20XX",
-      shortText: "Creativity to incorporate the paragraph into their writing",
-      fullText: "If you're looking for random paragraphs, you've come to the right place. When a random word or a random sentence isn't quite enough, the next logical step is to find a random paragraph. We created the Random Paragraph Generator with you in mind. The process is quite simple. Choose the number of random paragraphs you'd like to see and click the button. Your chosen number of paragraphs will instantly appear. While it may not be obvious to everyone, there are a number of reasons creating random paragraphs can be useful. A few examples of how some people use this generator are listed in the following paragraphs.",
-      tags: [
-        { name: "Records", url: "/" },
-        { name: "V3 model", url: "/bikes" },
-        { name: "Bike", url: "/bikes" },
-        { name: "Top Speed", url: "/bikes" },
-        { name: "Race", url: "/" }
-      ]
-    },
-    {
-      title: "Thrilling Trike Trials!",
-      image: "/trike_race2.jpg",
-      author: "Alvin Fung",
-      date: "XX-ABC-21XX",
-      shortText: "Creativity to incorporate the paragraph into their writing",
-      fullText: "t was the best compliment that he'd ever received although the person who gave it likely never knew. It had been an off-hand observation on his ability to hold a conversation and actually add pertinent information to it on practically any topic. Although he hadn't consciously strived to be able to do so, he'd started to voraciously read the news when he couldn't keep up on topics his friends discussed because their conversations went above his head. The fact that someone had noticed enough to compliment him that he could talk intelligently about many topics meant that he had succeeded in his quest to be better informed.",
-      tags: [
-        { name: "Vehicles", url: "/bikes" },
-        { name: "Race", url: "/" },
-        { name: "Trike", url: "/bikes" }
-      ]
-    },
-    {
-      title: "Battle Mountain Trials!",
-      image: "/battle_mountain_group.jpg",
-      author: "Tran Ngoc Duy Ngo",
-      date: "XX-ABC-21XX",
-      shortText: "Creativity to incorporate the paragraph into their writing",
-      fullText: "Generating random paragraphs can be an excellent way for writers to get their creative flow going at the beginning of the day. The writer has no idea what topic the random paragraph will be about when it appears. This forces the writer to use creativity to complete one of three common writing challenges. The writer can use the paragraph as the first one of a short story and build upon it. A second option is to use the random paragraph somewhere in a short story they create. The third option is to have the random paragraph be the ending paragraph in a short story. No matter which of these challenges is undertaken, the writer is forced to use creativity to incorporate the paragraph into their writing.",
-      tags: [
-        { name: "Mountain", url: "/" },
-        { name: "Race", url: "/" },
-        { name: "Trike", url: "/bikes" },
-        { name: "Competitions", url: "/" }
-      ]
-    },
-    {
-      title: "O-week Greetings!",
-      image: "/o_week_group.jpg",
-      author: "Tran Ngoc Duy Ngo",
-      date: "XX-ABC-21XX",
-      shortText: "Creativity to incorporate the paragraph into their writing",
-      fullText: "Generating random paragraphs can be an excellent way for writers to get their creative flow going at the beginning of the day. The writer has no idea what topic the random paragraph will be about when it appears. This forces the writer to use creativity to complete one of three common writing challenges. The writer can use the paragraph as the first one of a short story and build upon it. A second option is to use the random paragraph somewhere in a short story they create. The third option is to have the random paragraph be the ending paragraph in a short story. No matter which of these challenges is undertaken, the writer is forced to use creativity to incorporate the paragraph into their writing.",
-      tags: [
-        { name: "Introduction", url: "/join" },
-        { name: "Teams", url: "/" },
-        { name: "Trike", url: "/bikes" },
-        { name: "Operations", url: "/" }
-      ]
-    },
-    {
-      title: "Weekly Team Meeting!",
-      image: "/mhp_group.jpg",
-      author: "Tran Ngoc Duy Ngo",
-      date: "XX-ABC-21XX",
-      shortText: "creativity to incorporate the paragraph into their writing",
-      fullText: "Generating random paragraphs can be an excellent way for writers to get their creative flow going at the beginning of the day. The writer has no idea what topic the random paragraph will be about when it appears. This forces the writer to use creativity to complete one of three common writing challenges. The writer can use the paragraph as the first one of a short story and build upon it. A second option is to use the random paragraph somewhere in a short story they create. The third option is to have the random paragraph be the ending paragraph in a short story. No matter which of these challenges is undertaken, the writer is forced to use creativity to incorporate the paragraph into their writing.",
-      tags: [
-        { name: "Vehicles", url: "/bikes" },
-        { name: "Race", url: "/" },
-        { name: "Trike", url: "/bikes" },
-        { name: "Operations", url: "/" }
-      ]
-    },
-  ];
-  
-  export default articles;
-
